requirements.txt
groq
requests==2.31.0
python-dotenv==1.0.0
newsdataapi
streamlit==1.32.0